refactor: fix naming inconsistencies across packages
