+++ packages/access/src/permission/evaluator/types.ts
@@ -20,7 +20,7 @@ export interface IPermissionEvaluator {
 }
 
 export type PermissionEvaluatorOptions = {
-    repository?: IPermissionProvider,
+    provider?: IPermissionProvider,
     policyEngine?: IPolicyEngine,
     realmId?: string | null,
     clientId?: string | null
